Add the "lazy start" technique to initialization stage
Previously, if an instance was started in read_only mode,
it could throw an error when initializing the queue module,
like this: "Can't modify data because this instance is in
read-only mode".
To avoid such behavior the "lazy start" technique has been added:
The queue module is loaded immediately if the instance was
configured with read_only = false. Otherwise, a start is delayed
until the instance will be configured with read_only = false.

Closes #122