drivers: add validation of the space to create_space
If a space is creating with the "if_not_exist = true" option and a
space with such name already exists (and this is not a tube of
the required type), a conflict will be detected and an error will
be thrown.
This is necessary because we don't want to create additional indexes
for the existing space and repurpose it into the "tube".

Closes #62