add "shared" sessions to queue
Before the patch a connection to server was synonym of the queue
session. Now the session has a unique UUID (returned by the
"queue.identify()" method), and one session can have many
connections.
The session will be deleted (all session tasks will be released) after
"ttr" seconds have passed since the last connection was disconnected.
To connect to an existing session, call "queue.identificate(uuid)"
with the previously obtained UUID.

"ttr" in seconds - the time after which, if there is no active
connection in the session, it will be released with all its tasks.

Also, the "_queue_taken" internal space has been replaced with the
"_queue_taken_2" with a change format and used indexes (for better
performance).
The downgrade to previous version works correctly.

Closes of #85