add queue settings
The patch adds the ability to set the queue settings(by calling
"queue.cfg(opts)").
Now only one setting is available - "ttr"(time to release).
"ttr" in seconds - the time after which, if there is no active
connection in the session, it will be released with all its tasks.

Part of #85