session: full support for shared sessions
In order for the user to be able to finish servicing the task after
the queue switched from replica to master, the `_queue_inactive_session`
space was renamed to `_queue_shared_sessions` and had an `active` field
in it. When the `take` method is called, the session uuid is automatically
added to the `_queue_shared_sessions` space with `exp_time` set to zero
and `active` field set to true. When the client closes the session, this
entry becomes inactive (`active` = false) and `exp_time` is set.

At the start of the queue, we make all active sessions inactive, and set
an expiration time for them. If the `ttr` setting is not set, then we delete
all sessions.

If `ttr` is not set shared sessions is disabled.

Follows up #120