cfg: check for temporary tubes in replicaset mode
In replicaset mode we cannot use temporary tubes.
Data on them will not be replicated and the correct
operation of the queue will be disrupted.

This patch will not allow creation of the temporary tube
if `in_replicaset == true` or will not allow to set
`in_replicaset = true` if temporary tubes exists.

Follows up #120