abstract: reduced increased CPU consumption
Due to the fact that queue_state_fiber called box.ctrl.wait_r*
every millisecond, the idle queue consumed more than 10% of CPU.

The proposed solution makes calls to box.ctrl.wait_r* blocking
until the read/write mode is changed. As a result, in idle mode,
the queue_state_faber is in the suspended state, while the main
fiber (for example, interactive in interactive-mode) is in the
running state.
Thus, CPU consumption is reduced to about ~1%.

More detailed measurements are described in [1].

1. https://github.com/tarantool/queue/pull/192

Closes #183