replicaset_mode: allow api call in init state
In replicaset mode, the user can attempt to call public methods on the replica start.
For example, a single script is used for master and replica.

This patch allows you to call a public method in the INIT state without
interrupting the script by exiting with an obscure error.
It will only log an attempt to call a method from the check_state function,
which is present at the beginning of each public method.

Closes #216