core: make queue_state_fiber cancellable
We are going to finish all client (non-system) fibers in the process of
Tarantool shutdown. First we cancel them and then wait for their
finishing. So if the client fibers are not finished Tarantool shutdown
is never finished too.

Currently the fiber can not be finished and we got hang in integration
testing of PR tarantool/tarantool#9604.